Move DEBIAN_FRONTEND env var inline
DEBIAN_FRONTEND=noninteractive should not be set globally in the
environment because later apt-get installations may indeed be
interactive.

This patch moves the env var to be inline, so that it only affects the
command running in the Dockerfile.

The README was also updated to be more general for easier maintenance.